Move test to NxosBgpTest
